backend/requirements.txt
flask==2.0.3
werkzeug==2.0.3
flask-cors==3.0.10
psycopg2-binary==2.9.3
python-dotenv==0.19.2